requirements.txt
# Core dependencies (you already have these)
streamlit==1.43.2
PyPDF2==3.0.1
pandas==2.2.3

# LangChain
langchain==0.3
langchain-core==0.3.45
langchain-community>=0.3.0
langchain-text-splitters==0.3.6

# Vector store and embeddings
faiss-cpu==1.10.0
sentence-transformers>=2.2.0

# LLM and transformers
transformers>=4.30.0
torch>=2.0.0

# Optional multimodal and document processing
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
camelot-py>=0.11.0
tabula-py>=2.9.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.5

# Testing
pytest>=8.3.0
